refactor(management): split procedures.py into per-procedure package
procedures.py (6 procedures) → procedures/ package with one file per
procedure under family subdirs (nm/, dm/, ...) carrying the spec
prefix in the filename. Scaffold for ~160 future procedures across
KNX 03.05.02 and 03.05.03.

Pure code move: function bodies byte-identical. Public API unchanged
(re-exports preserve every existing import path). 25 management
tests pass unchanged.
