+++ docs/changelog.md
@@ -6,6 +6,12 @@ nav_order: 2
 
 # Changelog
 
+# Unreleased
+
+### Internals
+
+- Split `xknx/management/procedures.py` into the `xknx/management/procedures/` package. Each procedure lives in its own file under a family subdirectory (`nm/`, `dm/`, etc.) with the KNX spec prefix in the filename. Public API and behaviour unchanged.
+
 # 3.15.0 Task improvements 2026-02-15
 
 ### Telegram
