+++ test/management_tests/procedures/dm/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for KNX 03.05.02 DM_* procedures."""
