+++ test/management_tests/procedures/nm/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for KNX 03.05.02 NM_* procedures."""
