+++ xknx/management/procedures/dm/__init__.py
@@ -1,0 +1,4 @@
+"""KNX 03.05.02 Device Management (DM_*) procedures."""
+
+# ruff: noqa: F401
+from .dm_restart_rco import dm_restart
