+++ xknx/management/procedures/nm/__init__.py
@@ -1,0 +1,12 @@
+"""KNX 03.05.02 Network Management (NM_*) procedures."""
+
+# ruff: noqa: F401
+from .nm_individual_address_check import nm_individual_address_check
+from .nm_individual_address_read import nm_individual_address_read
+from .nm_individual_address_serial_number_read import (
+    nm_individual_address_serial_number_read,
+)
+from .nm_individual_address_serial_number_write import (
+    nm_individual_address_serial_number_write,
+)
+from .nm_individual_address_write import nm_individual_address_write
