+++ xknx/management/procedures/nm/nm_individual_address_check.py
@@ -1,0 +1,49 @@
+"""NM_IndividualAddress_Check — KNX 03.05.02 §2.19."""
+
+from __future__ import annotations
+
+import logging
+from typing import TYPE_CHECKING
+
+from xknx.exceptions import ManagementConnectionRefused, ManagementConnectionTimeout
+from xknx.telegram import apci
+from xknx.telegram.address import IndividualAddress, IndividualAddressableType
+
+if TYPE_CHECKING:
+    from xknx import XKNX
+
+logger = logging.getLogger("xknx.management.procedures")
+
+
+async def nm_individual_address_check(
+    xknx: XKNX, individual_address: IndividualAddressableType
+) -> bool:
+    """
+    Check if the individual address is occupied on the network.
+
+    :param xknx: XKNX object
+    :param individual_address: address to check
+    """
+    try:
+        async with xknx.management.connection(
+            address=IndividualAddress(individual_address)
+        ) as connection:
+            try:
+                response = await connection.request(
+                    payload=apci.DeviceDescriptorRead(descriptor=0),
+                    expected=apci.DeviceDescriptorResponse,
+                )
+
+            except ManagementConnectionTimeout as ex:
+                # if nothing is received (-> timeout) IA is free
+                logger.debug("No device answered to connection attempt. %s", ex)
+                return False
+            if isinstance(response.payload, apci.DeviceDescriptorResponse):
+                # if response is received IA is occupied
+                logger.debug("Device found at %s", individual_address)
+                return True
+            return False
+    except ManagementConnectionRefused as ex:
+        # if Disconnect is received immediately, IA is occupied
+        logger.debug("Device does not support transport layer connections. %s", ex)
+        return True
