+++ xknx/management/procedures/nm/nm_individual_address_serial_number_read.py
@@ -1,0 +1,35 @@
+"""NM_IndividualAddress_SerialNumber_Read — KNX 03.05.02 §2.4."""
+
+from __future__ import annotations
+
+import logging
+from typing import TYPE_CHECKING
+
+from xknx.telegram import apci
+from xknx.telegram.address import IndividualAddress
+
+if TYPE_CHECKING:
+    from xknx import XKNX
+
+logger = logging.getLogger("xknx.management.procedures")
+
+
+async def nm_individual_address_serial_number_read(
+    xknx: XKNX,
+    serial: bytes,
+    timeout: float = 3,
+) -> IndividualAddress | None:
+    """Read individual address from device with specified serial number."""
+    # initialize queue or event handler gathering broadcasts
+    async with xknx.management.broadcast() as bc_context:
+        await xknx.management.send_broadcast(
+            payload=apci.IndividualAddressSerialRead(serial=serial)
+        )
+        async for result in bc_context.receive(timeout=timeout):
+            if (
+                isinstance(result.payload, apci.IndividualAddressSerialResponse)
+                and result.payload.serial == serial
+            ):
+                return result.source_address
+
+    return None
